+++ hawc/apps/hawc_admin/urls.py
@@ -19,7 +19,6 @@ def get_admin_urlpatterns(open_api_patterns) -> List:
 
     admin_url = f"admin/{settings.ADMIN_URL_PREFIX}" if settings.ADMIN_URL_PREFIX else "admin"
 
-    # always include API for healthchecks
     router = DefaultRouter()
     if settings.INCLUDE_ADMIN:
         router.register(r"dashboard", api.DashboardViewset, basename="admin_dashboard")
