+++ tests/hawc/apps/common/test_api.py
@@ -1,15 +1,21 @@
 import json
 
 import pytest
+from django.conf import settings
 from django.urls import reverse
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.test import APIClient
 
 from hawc.apps.common.api import user_can_edit_object
+from hawc.apps.common.diagnostics import worker_healthcheck
 from hawc.apps.myuser.models import HAWCUser
 from hawc.apps.study.models import Study
 
 from ..test_utils import check_details_of_last_log_entry
+
+
+def has_redis():
+    return "RedisCache" in settings.CACHES["default"]["BACKEND"]
 
 
 @pytest.mark.django_db
@@ -164,3 +170,46 @@ def test_user_can_edit_object(db_keys):
     assert user_can_edit_object(final, user) is False
     with pytest.raises(PermissionDenied):
         user_can_edit_object(final, user, raise_exception=True)
+
+
+@pytest.mark.django_db
+class TestHealthcheckViewset:
+    def test_web(self):
+        client = APIClient()
+        url = reverse("healthcheck-web")
+        resp = client.get(url)
+        assert resp.status_code == 200
+        assert resp.json() == {"healthy": True}
+
+    @pytest.mark.skipif(not has_redis(), reason="skip; redis cache required")
+    def test_worker(self):
+        client = APIClient()
+        url = reverse("healthcheck-worker")
+
+        # no data; should be an error
+        resp = client.get(url)
+        assert resp.status_code == 400
+        assert resp.json()["healthy"] is False
+
+        # has recent data; should be healthy
+        worker_healthcheck.push()
+        resp = client.get(url)
+        assert resp.status_code == 200
+        assert resp.json()["healthy"] is True
+
+    @pytest.mark.skipif(not has_redis(), reason="skip; redis cache required")
+    def test_worker_plot(self):
+        client = APIClient()
+        url = reverse("healthcheck-worker-plot")
+
+        # failure - not admin
+        resp = client.get(url)
+        assert resp.status_code == 403
+        assert client.login(username="pm@hawcproject.org", password="pw") is True
+        resp = client.get(url)
+        assert resp.status_code == 403
+
+        # success - admin
+        assert client.login(username="admin@hawcproject.org", password="pw") is True
+        resp = client.get(url)
+        assert resp.status_code == 200
